+++ bubo/rooms.py
@@ -1,10 +1,16 @@
 import logging
 import time
 from copy import deepcopy
-from typing import Tuple, Optional, List, Dict
+from typing import Tuple, Optional, List, Dict, Union
 
+from aiohttp import ClientResponse
 # noinspection PyPackageRequirements
-from nio import AsyncClient, RoomVisibility, EnableEncryptionBuilder, RoomPutStateError
+from nio import (
+    AsyncClient, RoomVisibility, EnableEncryptionBuilder, RoomPutStateError, RoomGetStateEventError,
+    RoomPutStateResponse, RoomGetStateEventResponse,
+)
+# noinspection PyPackageRequirements
+from nio.http import TransportResponse
 
 from bubo.chat_functions import invite_to_room
 from bubo.config import Config
@@ -186,21 +192,39 @@ async def ensure_room_exists(
     return "exists", None
 
 
-async def set_user_power(room_id: str, user_id: str, client: AsyncClient, power: int):
+async def set_user_power(
+    room_id: str, user_id: str, client: AsyncClient, power: int,
+) -> Union[int, RoomGetStateEventError, RoomGetStateEventResponse, RoomPutStateError, RoomPutStateResponse]:
     """
     Set user power in a room.
     """
     logger.debug(f"Setting user power: {room_id}, user: {user_id}, level: {power}")
-    state = await client.room_get_state_event(room_id, "m.room.power_levels")
-    state.content["users"][user_id] = power
+    state_response = await client.room_get_state_event(room_id, "m.room.power_levels")
+    if isinstance(state_response, RoomGetStateEventError):
+        logger.error(f"Failed to fetch room {room_id} state: {state_response.message}")
+        return state_response
+    if isinstance(state_response.transport_response, TransportResponse):
+        status_code = state_response.transport_response.status_code
+    elif isinstance(state_response.transport_response, ClientResponse):
+        status_code = state_response.transport_response.status
+    else:
+        logger.error(f"Failed to determine status code from state response: {state_response}")
+        return state_response
+    if status_code >= 400:
+        logger.warning(
+            f"Failed to set user {user_id} power in {room_id}, response {status_code}"
+        )
+        return status_code
+    state_response.content["users"][user_id] = power
     response = await with_ratelimit(
         client,
         "room_put_state",
         room_id=room_id,
         event_type="m.room.power_levels",
-        content=state.content,
+        content=state_response.content,
     )
     logger.debug(f"Power levels update response: {response}")
+    return response
 
 
 async def maintain_configured_rooms(client: AsyncClient, store: Storage, config: Config):